Moved plot_implicit function to a new file.
