+++ sympy/plotting/__init__.py
@@ -1,4 +1,5 @@
-from plot import plot, plot_backends, plot_implicit
+from plot import plot, plot_backends
+from plot_implicit import plot_implicit
 from proxy_pyglet import Plot
 from textplot import textplot
 from pygletplot import PygletPlot
